从 webui.py 中拆分出 ui/state.py，顺便修复一个 windrecorder/state.py 中 gap_num 可能为 0 的 bug
